+++ ludo-server/src/main/java/app/repositories/UserLobbyRepository.java
@@ -24,13 +24,13 @@ import java.util.List;
         return namedQuery.getResultList();
     }
 
-    public List<UserLobby> findAllUsersInLobby(int id){
+    public List<UserLobby> findAllUsersInLobby(Long id){
         TypedQuery<UserLobby> namedQuery = entityManager.createNamedQuery("find_users_in_lobby", UserLobby.class);
         namedQuery.setParameter(1, id);
         return namedQuery.getResultList();
     }
 
-    public void removeUserFromLobby(String username, int lobbyId) {
+    public void removeUserFromLobby(String username, Long lobbyId) {
         Query namedQuery = entityManager.createNamedQuery("remove_user_from_lobby");
         namedQuery.setParameter("username", username);
         namedQuery.setParameter("lobbyId", lobbyId);
@@ -41,12 +41,12 @@ import java.util.List;
         return entityManager.merge(userLobby);
     }
 
-    public UserLobby findById(int id){
+    public UserLobby findById(Long id){
             return entityManager.find(UserLobby.class, id);
     }
 
 
-    public UserLobby deleteById(int id){
+    public UserLobby deleteById(Long id){
             UserLobby lobby = findById(id);
             entityManager.remove(lobby);
             return lobby;
